homecredit-model-3/requirements.txt
polars-lts-cpu==0.20.20
scikit-learn
numpy==1.26.4
pandas==2.2.2
gdown
lightgbm
catboost